FE-43: Consolidate filter pipeline into FilterShell with compositing prop
Remove rasterized displacement map, specular map, and all intermediate filter
variants (Filter, FilterOBB, FilterPolar, FilterPolarHiRes). The HOC now uses
FilterShell directly with a pre-computed hi-res geometric polar field and
SVG filter math (PolarToCartesian + magnitude lookup table).

- Extract calculateDisplacementMapRadius to maps/displacement-radius.ts
- Move composite components to components/composite/{image,parts}.tsx
- FilterShell is now the full pipeline: blur → compositing → PolarToCartesian → displacement
- Add compositing prop ("image" default, "parts") to choose compositing strategy
- Simplify calculateGeometricPolarMap to take only radius
- Remove specular support (to be re-implemented as SVG-only filter later)
